Bug fix: Issue #6
If there is punctuation in the utterance, e.g.:

I like pizza.
I like pizza!
Do you like pizza?
You like pizza, don’t you

The punctuation sticks to the word because the split is done with spaces only. Separating the punctuation actually breaks the grammar so the only hack that works is to remove the punctuation: |,.? from the utterance.

We need a proper tokeniser that creates the utterance in Chat before we pass it to text2triple module as a list of tokens.